+++ Deeploy/Targets/Snitch/Templates/RQAddTemplate.py
@@ -25,30 +25,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import Dict, List, Tuple
+from Deeploy.Targets.Generic.Templates.RQAddTemplate import RQAddTemplate
 
-from Deeploy.DeeployTypes import NetworkContext, NodeTemplate, OperatorRepresentation
-
-
-class SnitchRQAddTemplate(NodeTemplate):
-
-    def __init__(self, templateStr):
-        super().__init__(templateStr)
-
-    def alignToContext(self, ctxt: NetworkContext,
-                       operatorRepresentation: OperatorRepresentation) -> Tuple[NetworkContext, Dict, List[str]]:
-        # Extract signedness information of input, weights and output
-        signedI2 = ctxt.lookup(operatorRepresentation['data_in_2'])._type.referencedType.typeMin < 0
-        signedI = ctxt.lookup(operatorRepresentation['data_in_1'])._type.referencedType.typeMin < 0
-        signedO = ctxt.lookup(operatorRepresentation['data_out'])._type.referencedType.typeMin < 0
-        operatorRepresentation['input_2_signed'] = signedI2
-        operatorRepresentation['input_signed'] = signedI
-        operatorRepresentation['output_signed'] = signedO
-
-        return ctxt, operatorRepresentation, []
-
-
-referenceTemplate = SnitchRQAddTemplate("""
+referenceTemplate = RQAddTemplate("""
 
 <%
 signatureString = ''
@@ -67,5 +46,5 @@ else:
 %>
 
 // PULP NN RQADD
-pulp_nn_add${signatureString}(${data_in_1}, ${data_in_2}, ${data_out}, ${rqs1_mul}, ${rqs1_add}, ${rqs1_log2D}, ${rqs2_mul}, ${rqs2_add}, ${rqs2_log2D}, ${rqsOut_mul}, ${rqsOut_add}, ${rqsOut_log2D}, 1, ${size}, 1, 1);
+snitch_nn_add${signatureString}(${data_in_1}, ${data_in_2}, ${data_out}, ${rqs1_mul}, ${rqs1_add}, ${rqs1_log2D}, ${rqs2_mul}, ${rqs2_add}, ${rqs2_log2D}, ${rqsOut_mul}, ${rqsOut_add}, ${rqsOut_log2D}, 1, ${size}, 1, 1);
 """)
